fix error with finally

+++ API/Controllers/KeyVaultController.cs
@@ -395,9 +395,11 @@ public class KeyVaultController : Controller
         
         // Try to create an alert for the new key
         _logger.LogInformation("Creating a log alert for the new key {keyName}", request.Name);
+        bool created;
         try
         {
             await _alertService.CreateAlertForKeyAsync($"{request.Name}-Key-Alert", response.Key.Kid!, request.ActionGroups);
+            created = true;
         }
         catch (RequestFailedException e)
         {
@@ -418,7 +420,7 @@ public class KeyVaultController : Controller
             return StatusCode(StatusCodes.Status500InternalServerError, "An unexpected error occurred");
         }
         // Delete the new key if creating the alert fails
-        finally
+        if (!created)
         {
             await _keyVaultService.DeleteKeyAsync(request.Name);
             _logger.LogWarning("The key {keyName} was deleted because the alert could not be created", request.Name);
